initial tests cleared
