+++ tox.ini
@@ -188,7 +188,6 @@ exclude =
     wandb/sdk_py27/,
     wandb/proto/,
     wandb/util.py,
-    wandb/data_types.py,
     wandb/env.py,
     wandb/sweeps/,
     wandb/integration/magic.py,
